refactor: separate Claude message responsibilities
